@@ -24,10 +24,10 @@
                 <version>3.10.1</version>
                 <configuration>
                     <encoding>utf-8</encoding>
-                    <source>8</source>
-                    <target>8</target>
-                    <testSource>8</testSource>
-                    <testTarget>8</testTarget>
+                    <source>17</source>
+                    <target>17</target>
+                    <testSource>17</testSource>
+                    <testTarget>17</testTarget>
                 </configuration>
             </plugin>
         </plugins>
